if you were slick enough to figure out the swap file name ... which is not hard ".swapN" N corresponding to a number that gets incremented at every swap creation... however, it is hard to keep up because after a pcb has been swapped in, that is deleted and a new one is created for the next swap.
with that background said -- if you're fast enough, you could delete the
swap file..which would result in the program just terminating.

I added a check to the shell command to not allow any filenames that
start with a . to be deleted (per the command) delete in the device
driver can still delete it.

+++ source/scripts/os/shell.ts
@@ -853,6 +853,10 @@ module TSOS {
                 _StdOut.putText("Invalid file name");
                 return;
             }
+            if(args[0].charAt(0) === "."){
+                _StdOut.putText("Nice try, cannot delete filenames that start with '.', in fact how do you know they exists??");
+                return;
+            }
             // was the file deleted?
             var deleted = _krnHDDdriver.deleteFile(args[0]);
             // say it has been deleted
